1_zhYuan 1.enable HSIC interface (include VHSIC_1V2, GPIO102) 2.shell for auto load

+++ drivers/usb/host/comip/comip_hcd_ddma.c
@@ -597,7 +597,7 @@ static void init_non_isoc_dma_desc(comip_hcd_t * hcd, comip_qh_t * qh)
 
         if (n_desc) {
             /* SG request - more than 1 QTDs */
-            hc->xfer_buff = (uint8_t *)((unsigned int)qtd->urb->dma + qtd->urb->actual_length);
+            hc->xfer_buff = (uint8_t *)qtd->urb->dma + qtd->urb->actual_length;
             hc->xfer_len = qtd->urb->length - qtd->urb->actual_length;
         }
